+++ druid/base/hover.lua
@@ -5,10 +5,10 @@
 -- @within BaseComponent
 -- @alias druid.hover
 
---- On hover callback(self, state)
+--- On hover callback(self, state, hover_instance)
 -- @tfield DruidEvent on_hover @{DruidEvent}
 
---- On mouse hover callback(self, state)
+--- On mouse hover callback(self, state, hover_instance)
 -- @tfield DruidEvent on_mouse_hover @{DruidEvent}
 
 ---
@@ -81,6 +81,8 @@ function Hover.on_input(self, action_id, action)
 	else
 		hover_function(self, true)
 	end
+
+	return false
 end
 
 
@@ -95,9 +97,18 @@ end
 function Hover.set_hover(self, state)
 	if self._is_hovered ~= state then
 		self._is_hovered = state
-		self.on_hover:trigger(self:get_context(), state)
+		self.on_hover:trigger(self:get_context(), state, self)
 	end
 end
+
+
+--- Return current hover state. True if touch action was on the node at current time
+-- @tparam Hover self @{Hover}
+-- @treturn bool The current hovered state
+function Hover.is_hovered(self)
+	return self._is_hovered
+end
+
 
 --- Set mouse hover state
 -- @tparam Hover self @{Hover}
@@ -105,8 +116,16 @@ end
 function Hover.set_mouse_hover(self, state)
 	if self._is_mouse_hovered ~= state then
 		self._is_mouse_hovered = state
-		self.on_mouse_hover:trigger(self:get_context(), state)
+		self.on_mouse_hover:trigger(self:get_context(), state, self)
 	end
+end
+
+
+--- Return current hover state. True if nil action_id (usually desktop mouse) was on the node at current time
+-- @tparam Hover self @{Hover}
+-- @treturn bool The current hovered state
+function Hover.is_mouse_hovered(self)
+	return self._is_mouse_hovered
 end