+++ druid/helper.lua
@@ -12,7 +12,7 @@ local M = {}
 --- Text node or icon node can be nil
 local function get_text_width(text_node)
 	if text_node then
-		local text_metrics = gui.get_text_metrics_from_node(text_node)
+		local text_metrics = M.get_text_metrics_from_node(text_node)
 		local text_scale = gui.get_scale(text_node).x
 		return text_metrics.width * text_scale
 	end
@@ -107,6 +107,13 @@ function M.get_screen_aspect_koef()
 end
 
 
+function M.get_gui_scale()
+	local window_x, window_y = window.get_size()
+	return math.min(window_x / gui.get_width(),
+		window_y / gui.get_height())
+end
+
+
 function M.step(current, target, step)
 	if current < target then
 		return math.min(current + step, target)
@@ -166,6 +173,21 @@ function M.contains(t, value)
 end
 
 
+--- Get text metric from gui node. Replacement of previous gui.get_text_metrics_from_node function
+-- @tparam Node text_node
+-- @treturn table {width, height, max_ascent, max_descent}
+function M.get_text_metrics_from_node(text_node)
+	local font_name = gui.get_font(text_node)
+	local font = gui.get_font_resource(font_name)
+	return resource.get_text_metrics(font, gui.get_text(text_node), {
+		width = gui.get_size(text_node).x,
+		leading = gui.get_leading(text_node),
+		tracking = gui.get_tracking(text_node),
+		line_break = gui.get_line_break(text_node),
+	})
+end
+
+
 --- Check if node is enabled in gui hierarchy.
 -- Return false, if node or any his parent is disabled
 -- @function helper.is_enabled
@@ -180,6 +202,23 @@ function M.is_enabled(node)
 	end
 
 	return is_enabled
+end
+
+
+--- Get cumulative parent's node scale
+-- @function helper.get_scene_scale
+-- @tparam node node Gui node
+-- @tparam bool include_passed_node_scale True if add current node scale to result
+-- @treturn vector3 The scene node scale
+function M.get_scene_scale(node, include_passed_node_scale)
+	local scale = include_passed_node_scale and gui.get_scale(node) or vmath.vector3(1)
+	local parent = gui.get_parent(node)
+	while parent do
+		scale = vmath.mul_per_elem(scale, gui.get_scale(parent))
+		parent = gui.get_parent(parent)
+	end
+
+	return scale
 end
 
 
@@ -277,6 +316,26 @@ function M.get_border(node, offset)
 	end
 
 	return border
+end
+
+
+--- Get text metric from gui node. Replacement of previous gui.get_text_metrics_from_node function
+-- @tparam Node text_node
+-- @treturn table {width, height, max_ascent, max_descent}
+function M.get_text_metrics_from_node(text_node)
+	local font_resource = gui.get_font_resource(gui.get_font(text_node))
+	local options = {
+		tracking = gui.get_tracking(text_node),
+		line_break = gui.get_line_break(text_node),
+	}
+
+	-- Gather other options only if it used in node
+	if options.line_break then
+		options.width = gui.get_size(text_node).x
+		options.leading = gui.get_leading(text_node)
+	end
+
+	return resource.get_text_metrics(font_resource, gui.get_text(text_node), options)
 end