+++ druid/styles/default/style.lua
@@ -56,6 +56,7 @@ M["button"] = {
 
 M["drag"] = {
 	DRAG_DEADZONE = 10, -- Size in pixels of drag deadzone
+	NO_USE_SCREEN_KOEF = false,
 }
 
 
@@ -113,6 +114,7 @@ M["input"] = {
 	BUTTON_SELECT_INCREASE = 1.06,
 	MASK_DEFAULT_CHAR = "*",
 	IS_UNSELECT_ON_RESELECT = false,
+	NO_CONSUME_INPUT_WHILE_SELECTED = false,
 
 	on_select = function(self, button_node)
 		local target_scale = self.button.start_scale